+++ api-scanner/src/test/resources/kotlin-legacy/kotlin/net/corda/example/LegacyApi.kt
@@ -1,0 +1,13 @@
+package net.corda.example
+
+import net.corda.annotation.AnAnnotation
+
+@AnAnnotation
+class LegacyApi {
+    init {
+        this::class.annotations.forEach {
+            println(it.javaClass.name)
+        }
+    }
+}
+